mac_requirements.txt
biopython==1.85
numpy==2.3.1
py2app